Can successfully sort integer sequences

+++ examples/set2seq/utils/python_utils.py
@@ -29,4 +29,4 @@ def read_data(txt_data):
   data = open_txt(txt_data)
   data = [d.split(' ') for d in data]
   data = [[int(d) for d in data_list] for data_list in data]
-  return np.array(data)
+  return np.array(data).astype(int)
